ApplicationDbContext.cs
EmployeeLicenseType.cs
EmployeeLicenseTypeAudit.cs
EmployeeLicensure.cs
TaskAudit.cs
TaskController.cs
EmployeeLicenseTypeViewModel
EmployeeLicensureViewModel.cs

EmployeeLicenseType changed Data type for  EmployeeLicenseTypeId
EmployeeLicenseTypeAudit changed Data type for  EmployeeLicenseTypeId
EmployeeLicensure  changed Data type
Task Audit changed Id
Task Controller changed code
EmployeeLicensureViewModel changed data type

+++ DAL/Models/EmployeeLicenseTypeAudit.cs
@@ -9,7 +9,7 @@ namespace DAL.Models
     {
         [Key]
         public long EmployeeLicenseTypeAuditId { get; set; }
-        public Byte EmployeeLicenseTypeId { get; set; }
+        public long EmployeeLicenseTypeId { get; set; }
         public string Description { get; set; }
         public bool? IsActive { get; set; }
         public string CreatedBy { get; set; }
